requirements.txt
flask
flask-restplus==0.9.2
Flask-Cors==3.0.7
flask-swagger-ui==3.20.9
requests
#tensorflow==1.0.0
#sugartensor==1.0.0.2
#pandas==0.19.2
#librosa==0.6.2
#scikits.audiolab==0.11.0
python-socketio==4.4.0
gTTS
#pyttsx3
sanic
Sanic-Cors
pyaudio
SpeechRecognition
